Add tests for ModalClose controller
Amend tests for Confirm controller
